added script to create clean project which you can build with cpm.
This script is generic and also supports environments,
however there is a bug in cpm which does not allow you to build
with environments.

The general idea is to choose a specific environment and optional
add some extra libraries to it, instead of specifying all libraries
yourself. However because that environments is not working with cpm
we use the default environmnet of StdEnv which is the default for
cpm and add all needed systems libraries.

Added environment for linux/mac in IDEEnvs, because for those
operating systems Clean does not come with this file in its installation.
You have to copy this file into clean/etc/ folder to use it,
however as said before then building does not work anymore.